Emit `callFunction` with operators
The event `callFunction` was only being emitted for functions.
This commit fires this event when 'functions' are called with infix
operators. It also adds a test for that event firing correctly.

Catch function errors and emit `functionError` to supply another result

Add `prepare` to the list of scripts (called by Npm.depends in Meteor).